[Refactor] Rename ErrorPage to NotFoundPage
