+++ src/NotFoundPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function ErrorPage() {
+export default function NotFoundPage() {
   return (
     <h1>존재하지 않는 페이지 입니다</h1>
   );
