+++ src/NotFoundPage.test.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import ErrorPage from './ErrorPage';
+import NotFoundPage from './NotFoundPage';
 
-test('ErrorPage', () => {
+test('NotFoundPage', () => {
   const { queryByText } = render((
-    <ErrorPage />
+    <NotFoundPage />
   ));
 
   expect(queryByText('존재하지 않는 페이지 입니다')).not.toBeNull();
